feat: something

+++ server/src/main/java/com/pighand/aio/service/ECommerce/impl/ThemeServiceImpl.java
@@ -4,10 +4,15 @@ import com.mybatisflex.core.query.QueryWrapper;
 import com.pighand.aio.domain.ECommerce.ThemeDomain;
 import com.pighand.aio.mapper.ECommerce.ThemeMapper;
 import com.pighand.aio.service.ECommerce.ThemeService;
+import com.pighand.aio.service.common.UploadService;
 import com.pighand.aio.vo.ECommerce.ThemeVO;
 import com.pighand.framework.spring.base.BaseServiceImpl;
+import com.pighand.framework.spring.exception.ThrowException;
 import com.pighand.framework.spring.page.PageOrList;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 import static com.pighand.aio.domain.ECommerce.table.SessionTableDef.SESSION;
 import static com.pighand.aio.domain.ECommerce.table.SessionTemplateTableDef.SESSION_TEMPLATE;
@@ -20,7 +25,10 @@ import static com.pighand.aio.domain.ECommerce.table.ThemeTableDef.THEME;
  * @createDate 2024-05-23 15:01:58
  */
 @Service
+@RequiredArgsConstructor
 public class ThemeServiceImpl extends BaseServiceImpl<ThemeMapper, ThemeDomain> implements ThemeService {
+
+    private final UploadService uploadService;
 
     /**
      * 创建
@@ -32,6 +40,8 @@ public class ThemeServiceImpl extends BaseServiceImpl<ThemeMapper, ThemeDomain> 
     public ThemeVO create(ThemeVO themeVO) {
         super.mapper.insert(themeVO);
 
+        uploadService.updateFileOfficial(themeVO.getPosterUrl(), themeVO.getServiceQrUrl());
+
         return themeVO;
     }
 
@@ -42,8 +52,12 @@ public class ThemeServiceImpl extends BaseServiceImpl<ThemeMapper, ThemeDomain> 
      * @return
      */
     @Override
-    public ThemeDomain find(Long id) {
-        return super.mapper.find(id, SESSION.getTableName(), SESSION_TEMPLATE.getTableName());
+    public ThemeVO find(Long id) {
+        ThemeVO theme = super.mapper.find(id, SESSION.getTableName(), SESSION_TEMPLATE.getTableName());
+
+        // TODO: 无客服二维码，使用门店客服二维码
+
+        return theme;
     }
 
     /**
@@ -79,6 +93,29 @@ public class ThemeServiceImpl extends BaseServiceImpl<ThemeMapper, ThemeDomain> 
      */
     @Override
     public void update(ThemeVO themeVO) {
+        ThemeDomain themeDomain = super.queryChain().select(THEME.POSTER_URL).where(THEME.ID.eq(themeVO.getId())).one();
+
+        if (themeDomain == null) {
+            throw new ThrowException("未找到主题");
+        }
+
+        uploadService.replaceFileOfficial(themeDomain.getPosterUrl(), themeVO.getPosterUrl(), themeVO.getServiceQrUrl(),
+            themeDomain.getServiceQrUrl());
+
+        super.mapper.update(themeVO);
+    }
+
+    /**
+     * 修改排队时长
+     *
+     * @param id
+     * @param queueDuration
+     */
+    @Override
+    public void updateQueueDuration(Long id, Integer queueDuration) {
+        ThemeVO themeVO = new ThemeVO();
+        themeVO.setId(id);
+        themeVO.setQueueDuration(Optional.ofNullable(queueDuration).orElse(0));
         super.mapper.update(themeVO);
     }
 
@@ -89,6 +126,15 @@ public class ThemeServiceImpl extends BaseServiceImpl<ThemeMapper, ThemeDomain> 
      */
     @Override
     public void delete(Long id) {
+        ThemeDomain themeDomain = super.queryChain().select(THEME.POSTER_URL).where(THEME.ID.eq(id)).one();
+
+        if (themeDomain == null) {
+            throw new ThrowException("主题不存在");
+        }
+
         super.mapper.deleteById(id);
+
+        // 删除COS文件
+        uploadService.deleteFileOfficial(themeDomain.getPosterUrl(), themeDomain.getServiceQrUrl());
     }
 }
